+++ macros/src/codegen.rs
@@ -43,10 +43,7 @@ pub(crate) fn generate(args: Args, emitter: &mut Emitter) -> manyhow::Result {
         }
 
         rust_options.introspection_if = args.introspection_if.as_deref();
-
-        if let Some(ref krate) = args.krate {
-            rust_options.krate = krate;
-        }
+        rust_options.krate = args.krate.as_deref();
 
         let output = match generator.generate_rust(&rust_options) {
             Ok(output) => output,